refactor: MovieHeader에 선언되어 있던 SearchBox의 역할 위임
